Исправляет порционный показ комментариев
