+++ js/gallery-mini-photos.js
@@ -1,6 +1,6 @@
 import {photosDescription} from './data.js';
 import {openBigPhoto} from './modal-photo.js';
-// Отрисуйте сгенерированные DOM-элементы в блок .pictures.
+
 const photoListElements = document.querySelector('.pictures');
 const photoTemplate = document.querySelector('#picture').
   content.
@@ -32,5 +32,3 @@ const renderPhotoElements = (photos) => {
 };
 
 renderPhotoElements(photosDescription);
-//export {photoListElements};
-
